pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-security-patch"
version = "0.1.0"
description = "AI-powered security vulnerability remediation pipeline"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.42.0",
    "pydantic>=2.0",
    "PyGithub>=2.1",
    "gitpython>=3.1",
    "httpx>=0.27",
    "duckduckgo-search>=7.0",
    "beautifulsoup4>=4.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[tool.setuptools.packages.find]
include = ["pipeline*"]

[tool.setuptools.package-data]
pipeline = ["../agents/*.md"]
